pyproject.toml
[project]
name = "backend-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.21",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "clerk-backend-api>=4.2.0",
    "django>=5.2.9",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "django-stubs>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular>=0.29.0",
    "groq>=1.0.0",
    "huey>=2.5.5",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langchain-openai>=1.1.6",
    "markdown>=3.10",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-jose[cryptography]>=3.3.0",
    "redis>=7.1.0",
    "requests>=2.31.0",
    "sentry-sdk[django]>=2.49.0",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]
